Add activity update and delete endpoints and refactoring the test

See merge request inm5151-h24/touristes-de-salon-expires!109

+++ backend/__tests__/routes/activity.test.ts
@@ -7,6 +7,7 @@ import {Request, Response} from 'express';
 import * as actions from '../../src/services/user.services';
 import {activities} from '../../src/models/activities';
 import {closeDbConnection, db} from '../../src/db/db';
+import {eq} from 'drizzle-orm';
 
 const user = {username: 'test-user', password: '1234'};
 const activity = {
@@ -34,16 +35,16 @@ app.use(route, verifyUserToken, (req: Request, res: Response) => {
 jest.mock('../../src/services/user.services');
 
 beforeAll(async () => {
-  await db.delete(activities);
-  await db.delete(users);
+  await db.delete(activities).where(eq(activities.user_id, 1));
+  await db.delete(users).where(eq(users.id, 1));
 
   const hashedPassword = await bcrypt.hash(user.password, 10);
   returnedUser = {id: 1, username: user.username, password: hashedPassword};
 });
 
 afterAll(async () => {
-  await db.delete(activities);
-  await db.delete(users);
+  await db.delete(activities).where(eq(activities.user_id, 1));
+  await db.delete(users).where(eq(users.id, 1));
   return closeDbConnection();
 });
 
@@ -51,15 +52,14 @@ afterAll(async () => {
 describe('POST activity', () => {
   jest.spyOn(actions, 'getUserByUsername').mockImplementation(() => Promise.resolve(returnedUser));
   const route_creation : string = '/activity/manual';
+  const route_reception : string = '/activity/getActivity';
 
-  test('Should create a new activity', async () => {
+  test('#1 Should create a new activity', async () => {
     getToken = await request(app)
       .post('/auth')
       .send(user)
       .set('Content-Type', 'application/json');
-
     const { token } =  getToken.body;
-
 
     const res = await request(app)
       .post(route_creation)
@@ -69,7 +69,7 @@ describe('POST activity', () => {
     expect(res.statusCode).toBe(201);
   });
 
-  test('Should not work because not authenticated', async () => {
+  test('#2 Should not work because not authenticated', async () => {
     const res = await request(app)
       .post(route_creation)
       .send(activity);
@@ -77,7 +77,7 @@ describe('POST activity', () => {
     expect(res.statusCode).toBe(401);
   });
 
-  test('Should not create because of name', async () => {
+  test('#3 Should not create because of name', async () => {
     const activity = {
       name: 'this is a text of 257 characters: Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed eget turpis eu ligula ultricies fringilla. Nullam sed varius dui. Curabitur euismod, est ac venenatis eleifend, sapien purus varius justo, eget varius velit elit.',
       city: 'San Francisco',
@@ -91,9 +91,7 @@ describe('POST activity', () => {
       .post('/auth')
       .send(user)
       .set('Content-Type', 'application/json');
-
     const { token } =  getToken.body;
-
 
     const res = await request(app)
       .post(route_creation)
@@ -103,7 +101,7 @@ describe('POST activity', () => {
     expect(res.statusCode).toBe(400);
   });
 
-  test('Should not create because of type', async () => {
+  test('#4 Should not create because of type', async () => {
     const activity = {
       name: 'test',
       city: 'San Francisco',
@@ -117,20 +115,17 @@ describe('POST activity', () => {
       .post('/auth')
       .send(user)
       .set('Content-Type', 'application/json');
-
     const { token } =  getToken.body;
-
 
     const res = await request(app)
       .post(route_creation)
       .set('Authorization', `Bearer ${token}`)
       .send(activity);
-    console.log(res.text);
     expect(res.text).toEqual('{"message":"Type is required and must be one of the following: Running, Biking, Walking"}');
     expect(res.statusCode).toBe(400);
   });
 
-  test('Should not create because of date', async () => {
+  test('#5 Should not create because of date', async () => {
     const activity = {
       name: 'test',
       city: 'San Francisco',
@@ -144,9 +139,7 @@ describe('POST activity', () => {
       .post('/auth')
       .send(user)
       .set('Content-Type', 'application/json');
-
     const { token } =  getToken.body;
-
 
     const res = await request(app)
       .post(route_creation)
@@ -155,7 +148,7 @@ describe('POST activity', () => {
     expect(res.statusCode).toBe(400);
   });
 
-  test('Should not create because of durationTotal', async () => {
+  test('#6 Should not create because of durationTotal', async () => {
     const activity = {
       name: 'test',
       city: 'San Francisco',
@@ -169,22 +162,74 @@ describe('POST activity', () => {
       .post('/auth')
       .send(user)
       .set('Content-Type', 'application/json');
-
     const { token } =  getToken.body;
-
 
     const res = await request(app)
       .post(route_creation)
       .set('Authorization', `Bearer ${token}`)
       .send(activity);
-    expect(res.text).toEqual('{"message":"DurationTotal is required and must be a non-negative number"}');
+    expect(res.text).toEqual('{"message":"DurationTotal must be a non-negative and non-null number"}');
     expect(res.statusCode).toBe(400);
   });
 
+  test('#7 Should recover all activities information in the database', async () => {
+    jest.spyOn(actions, 'getUserById').mockImplementationOnce(() => Promise.resolve(returnedUser));
+    getToken = await request(app)
+      .post('/auth')
+      .send(user)
+      .set('Content-Type', 'application/json');
+    const { token } =  getToken.body;
+
+    const res = await request(app)
+      .get(route_reception)
+      .set('Authorization', `Bearer ${token}`);
+
+    expect(res.body.userActivities.length).toEqual(1);
+  });
+
+  test('#8 Should create a new activity', async () => {
+    const activity = {
+      name: 'test #2',
+      city: 'San Francisco',
+      type: 'Running',
+      date: '2024-02-28 15:45:00.123456',
+      durationTotal: 30,
+      distanceTotal: 7.5,
+      comment: 'Felt good!',
+    };
+
+    getToken = await request(app)
+      .post('/auth')
+      .send(user)
+      .set('Content-Type', 'application/json');
+    const { token } =  getToken.body;
+
+    const res = await request(app)
+      .post(route_creation)
+      .set('Authorization', `Bearer ${token}`)
+      .send(activity);
+
+    expect(res.statusCode).toBe(201);
+  });
+
+  test('#9 Should recover all activities information in the database', async () => {
+    jest.spyOn(actions, 'getUserById').mockImplementationOnce(() => Promise.resolve(returnedUser));
+    getToken = await request(app)
+      .post('/auth')
+      .send(user)
+      .set('Content-Type', 'application/json');
+    const { token } =  getToken.body;
+
+    const res = await request(app)
+      .get(route_reception)
+      .set('Authorization', `Bearer ${token}`);
+
+    expect(res.body.userActivities.length).toEqual(2);
+  });
 });
 
 describe('get all activities', () => {
-  const route_creation : string = '/activity/getActivity';
+  const route_reception : string = '/activity/getActivity';
 
   test('should recover all activities information in the database', async () => {
     jest.spyOn(actions, 'getUserById').mockImplementationOnce(() => Promise.resolve(returnedUser));
@@ -192,40 +237,91 @@ describe('get all activities', () => {
       .post('/auth')
       .send(user)
       .set('Content-Type', 'application/json');
-
     const { token } =  getToken.body;
 
-
     const res = await request(app)
-      .get(route_creation)
+      .get(route_reception)
       .set('Authorization', `Bearer ${token}`);
 
-    expect(res.statusCode).toEqual(200);
+    expect(res.body.userActivities.length).toEqual(2);
   });
-
 });
 
 describe('get specified activities', () => {
-  const route_creation : string = '/activity/getSpecifiedActivities';
+  const route_creation : string = '/activity/manual';
+  const route_reception_unique : string = '/activity/getSpecifiedActivities';
 
-
-  test('should recover searched activities', async () => {
+  test('#1 Should recover searched activities', async () => {
     jest.spyOn(actions, 'getUserById').mockImplementationOnce(() => Promise.resolve(returnedUser));
 
     getToken = await request(app)
       .post('/auth')
       .send(user)
       .set('Content-Type', 'application/json');
-
     const { token } =  getToken.body;
 
     const res = await request(app)
-      .get(`${route_creation}?search=Morning`)
+      .get(`${route_reception_unique}?search=Morning`)
       .set('Authorization', `Bearer ${token}`);
 
-    expect(res.statusCode).toEqual(200);
+    expect(res.body.activities.length).toEqual(0);
   });
 
+  test('#2 Should create a new activity', async () => {
+    getToken = await request(app)
+      .post('/auth')
+      .send(user)
+      .set('Content-Type', 'application/json');
+    const { token } =  getToken.body;
+
+    const res = await request(app)
+      .post(route_creation)
+      .set('Authorization', `Bearer ${token}`)
+      .send(activity);
+
+    expect(res.statusCode).toBe(201);
+  });
+
+  test('#3 Should create a new activity', async () => {
+    const activity = {
+      name: 'Morning',
+      city: 'San Francisco',
+      type: 'Running',
+      date: '2024-02-28 15:45:00.123456',
+      durationTotal: 30,
+      distanceTotal: 7.5,
+      comment: 'Felt good!',
+    };
+
+    getToken = await request(app)
+      .post('/auth')
+      .send(user)
+      .set('Content-Type', 'application/json');
+    const { token } =  getToken.body;
+
+    const res = await request(app)
+      .post(route_creation)
+      .set('Authorization', `Bearer ${token}`)
+      .send(activity);
+
+    expect(res.statusCode).toBe(201);
+  });
+
+  test('#4 Should recover searched activities', async () => {
+    jest.spyOn(actions, 'getUserById').mockImplementationOnce(() => Promise.resolve(returnedUser));
+
+    getToken = await request(app)
+      .post('/auth')
+      .send(user)
+      .set('Content-Type', 'application/json');
+    const { token } =  getToken.body;
+
+    const res = await request(app)
+      .get(`${route_reception_unique}?search=Morning`)
+      .set('Authorization', `Bearer ${token}`);
+
+    expect(res.body.activities.length).toEqual(1);
+  });
 });
 
 describe('GPX file related test', () => {
@@ -267,38 +363,89 @@ describe('GPX file related test', () => {
   });
 });
 
+describe('Delete activity', () => {
+  const route_deletion : string = '/activity/suppression/';
+  const route_reception : string = '/activity/getActivity';
+  const route_creation : string = '/activity/manual';
+
+  test('#1 Should create a new activity', async () => {
+    getToken = await request(app)
+      .post('/auth')
+      .send(user)
+      .set('Content-Type', 'application/json');
+    const { token } =  getToken.body;
+
+    const res = await request(app)
+      .post(route_creation)
+      .set('Authorization', `Bearer ${token}`)
+      .send(activity);
+
+    expect(res.statusCode).toBe(201);
+  });
+
+  test('#2 Should recover all activities information in the database', async () => {
+    jest.spyOn(actions, 'getUserById').mockImplementationOnce(() => Promise.resolve(returnedUser));
+    getToken = await request(app)
+      .post('/auth')
+      .send(user)
+      .set('Content-Type', 'application/json');
+    const { token } =  getToken.body;
+
+    const res = await request(app)
+      .get(route_reception)
+      .set('Authorization', `Bearer ${token}`);
+
+    expect(res.body.userActivities.length).toBeGreaterThan(0);
+  });
+
+  test('#3 Should delete every activity of the test user', async () => {
+    jest.spyOn(actions, 'getUserById').mockImplementationOnce(() => Promise.resolve(returnedUser));
+
+    getToken = await request(app)
+      .post('/auth')
+      .send(user)
+      .set('Content-Type', 'application/json');
+    const { token } =  getToken.body;
+
+    const res_request = await request(app)
+      .get(route_reception)
+      .set('Authorization', `Bearer ${token}`);
+
+
+    for (let i = 0; i < res_request.body.userActivities.length - 1; i++) {
+      await request(app)
+        .delete(route_deletion + res_request.body.userActivities[i].id)
+        .set('Authorization', `Bearer ${token}`);
+    }
+
+    const res = await request(app)
+      .delete(route_deletion + res_request.body.userActivities[res_request.body.userActivities.length - 1].id)
+      .set('Authorization', `Bearer ${token}`);
+
+    expect(res.statusCode).toEqual(200);
+  });
+
+  test('#4 Should recover all activities information in the database', async () => {
+    jest.spyOn(actions, 'getUserById').mockImplementationOnce(() => Promise.resolve(returnedUser));
+    getToken = await request(app)
+      .post('/auth')
+      .send(user)
+      .set('Content-Type', 'application/json');
+    const { token } =  getToken.body;
+
+    const res = await request(app)
+      .get(route_reception)
+      .set('Authorization', `Bearer ${token}`);
+
+    expect(res.body.userActivities.length).toEqual(0);
+  });
+
+});
 
 
 
 
 
-/**
- * This variable represents a GPX file in XML format.
- *
- * @type {string}
- * @description The GPX file contains location data of a track, including latitude, longitude, elevation, and time.
- * @example
- * gpxFileWorking = `<?xml version="1.0" encoding="UTF-8"?>
- * <gpx version="1.1" creator="AI Assistant">
- *   <metadata>
- *     <name>Example GPX</name>
- *     <desc>This is an example GPX file</desc>
- *   </metadata>
- *   <trk>
- *     <name>Example track</name>
- *     <trkseg>
- *       <trkpt lat="47.644548" lon="-122.326897">
- *         <ele>4.46</ele>
- *         <time>2009-10-17T18:36:26Z</time>
- *       </trkpt>
- *       <trkpt lat="47.644548" lon="-122.326897">
- *         <ele>4.94</ele>
- *         <time>2009-10-17T18:37:26Z</time>
- *       </trkpt>
- *     </trkseg>
- *   </trk>
- * </gpx>`;
- */
 export const gpxFileWorking = `<?xml version="1.0" encoding="UTF-8"?>
 <gpx version="1.1" creator="AI Assistant">
   <metadata>
@@ -320,33 +467,6 @@ export const gpxFileWorking = `<?xml version="1.0" encoding="UTF-8"?>
   </trk>
 </gpx>`;
 
-/**
- * @typedef {string} gpxFileWrong
- *
- * @description
- * The `gpxFileWrong` variable represents an invalid GPX file in XML format.
- *
- * @example
- * <?xml version="1.0" encoding="UTF-8"?>
- * <gpx version="1.1" creator="AI Assistant">
- *   <metadata>
- *     <name>Example GPX</name>
- *     <desc>This is an example GPX file</desc>
- *   </metadata>
- *   <trk>
- *     <name>Example track</name>
- *     <trkseg>
- *       <trkpt lat="47.644548" lon="-122.326897">
- *         <ele>4.46</ele>
- *         <time>2009-10-17T18:36:26Z</time>
- *       </trkpt>
- *       <trkpt lat="47.644548" lon="-122.326897">
- *         <ele>4.94</ele>
- *         <time>2009-10-17T18:37:26Z</time>
- *       </trkpt>
- *     </trkseg>
- *   </trk>
- */
 export const gpxFileWrong = `<?xml version="1.0" encoding="UTF-8"?>
 <gpx version="1.1" creator="AI Assistant">
   <metadata>
@@ -366,4 +486,3 @@ export const gpxFileWrong = `<?xml version="1.0" encoding="UTF-8"?>
       </trkpt>
     </trkseg>
   </trk>`;
-
